Move split tender tests to separate source file.
